feat: add DialogTogglePreferencesComponent and integrate it into overview and faculty preference components
