+++ frontend/src/app/core/components/admin/faculty-pref/faculty-pref.component.ts
@@ -19,7 +19,7 @@ import { TableHeaderComponent, InputField } from '../../../../shared/table-heade
 import { LoadingComponent } from '../../../../shared/loading/loading.component';
 import { DialogPrefComponent } from '../../../../shared/dialog-pref/dialog-pref.component';
 import { DialogExportComponent } from '../../../../shared/dialog-export/dialog-export.component';
-import { DialogActionComponent, DialogActionData } from '../../../../shared/dialog-action/dialog-action.component';
+import { DialogTogglePreferencesComponent, DialogTogglePreferencesData } from '../../../../shared/dialog-toggle-preferences/dialog-toggle-preferences.component';
 
 import { PreferencesService, ActiveSemester } from '../../../services/faculty/preference/preferences.service';
 
@@ -335,20 +335,17 @@ export class FacultyPrefComponent implements OnInit, AfterViewInit {
       ? new Date(this.allData[0].active_semesters[0].global_start_date)
       : null;
 
-    const hasIndividualDeadlines = this.hasIndividualDeadlines;
-
-    const dialogData: DialogActionData = {
+    const dialogData: DialogTogglePreferencesData = {
       type: 'all_preferences',
       academicYear: this.allData[0]?.active_semesters?.[0]?.academic_year || '',
       semester: this.allData[0]?.active_semesters?.[0]?.semester_label || '',
       currentState: this.isToggleAllChecked,
-      hasSecondaryText: false,
       global_deadline: existingDeadline,
       global_start_date: existingStartDate,
-      hasIndividualDeadlines: hasIndividualDeadlines,
+      hasIndividualDeadlines: this.hasIndividualDeadlines,
     };
 
-    const dialogRef = this.dialog.open(DialogActionComponent, {
+    const dialogRef = this.dialog.open(DialogTogglePreferencesComponent, {
       data: dialogData,
       disableClose: true,
       autoFocus: false,
@@ -356,13 +353,24 @@ export class FacultyPrefComponent implements OnInit, AfterViewInit {
 
     dialogRef.afterClosed().subscribe((confirmed: boolean) => {
       if (confirmed) {
-        const newStatus = this.isToggleAllChecked;
-        this.filteredData.forEach(
-          (faculty) => (faculty.is_enabled = newStatus)
-        );
-        this.isToggleAllChecked = newStatus;
-        this.updateDisplayedData();
-        this.cdr.markForCheck();
+        this.preferencesService.getPreferences().subscribe((response) => {
+          this.allData = response.preferences.map((faculty: any) => ({
+            faculty_id: faculty.faculty_id,
+            facultyName: faculty.faculty_name,
+            facultyCode: faculty.faculty_code,
+            facultyType: faculty.faculty_type,
+            facultyUnits: faculty.faculty_units,
+            has_request: faculty.has_request,
+            is_enabled: faculty.is_enabled === 1,
+            active_semesters: faculty.active_semesters,
+          }));
+          this.filteredData = [...this.allData];
+          this.applyFilter(this.currentFilter);
+          this.checkToggleAllState();
+          this.cdr.markForCheck();
+        });
+      } else {
+        this.isToggleAllChecked = !this.isToggleAllChecked;
       }
     });
   }
@@ -388,19 +396,18 @@ export class FacultyPrefComponent implements OnInit, AfterViewInit {
       activeSemester?.global_deadline ||
       null;
 
-    const dialogData: DialogActionData = {
+    const dialogData: DialogTogglePreferencesData = {
       type: 'single_preferences',
       academicYear: activeSemester?.academic_year || '',
       semester: activeSemester?.semester_label || '',
       currentState: faculty.is_enabled,
-      hasSecondaryText: false,
       facultyName: faculty.facultyName,
       faculty_id: faculty.faculty_id,
       individual_start_date: existingStartDate,
       individual_deadline: existingDeadline,
     };
 
-    const dialogRef = this.dialog.open(DialogActionComponent, {
+    const dialogRef = this.dialog.open(DialogTogglePreferencesComponent, {
       data: dialogData,
       disableClose: true,
       autoFocus: false,
@@ -408,20 +415,20 @@ export class FacultyPrefComponent implements OnInit, AfterViewInit {
 
     dialogRef.afterClosed().subscribe((confirmed: boolean) => {
       if (confirmed) {
-        faculty.is_enabled = !faculty.is_enabled;
-
         this.preferencesService.getPreferences().subscribe((response) => {
           const updatedFaculty = response?.preferences?.find(
             (item: any) => item.faculty_id === faculty.faculty_id
           );
           if (updatedFaculty) {
+            faculty.is_enabled = updatedFaculty.is_enabled === 1;
             faculty.active_semesters = updatedFaculty.active_semesters;
+            this.updateDisplayedData();
+            this.checkToggleAllState();
+            this.cdr.markForCheck();
           }
-
-          this.updateDisplayedData();
-          this.checkToggleAllState();
-          this.cdr.markForCheck();
         });
+      } else {
+        event.source.checked = faculty.is_enabled;
       }
     });
   }
