+++ frontend/src/app/shared/dialog-toggle-preferences/dialog-toggle-preferences.component.html
@@ -1,0 +1,201 @@
+<div class="dialog-action-container">
+  <!-- Dialog Action Title -->
+  <h2 class="dialog-title">{{ dialogTitle }}</h2>
+
+  <!-- Dialog Content -->
+  <div class="dialog-content">
+    <!-- Main Action Text -->
+    @if (!isPreferencesScheduled) {
+    <div class="main-text">
+      <strong>{{ actionText }}</strong>
+
+      <span>
+        {{
+          data.type === "all_preferences"
+            ? "load and schedule preferences submission for all active faculty members this"
+            : "load and schedule preferences submission for"
+        }}
+
+        {{ data.type.includes("single") ? data.facultyName + " this" : "" }}
+
+        A.Y. {{ data.academicYear }}, {{ data.semester }}.
+      </span>
+    </div>
+    }
+
+    <!-- Start Date Picker Section -->
+    @if (showStartDatePicker && !isPreferencesScheduled) {
+    <div class="date-picker">
+      <mat-form-field appearance="fill">
+        <mat-label>Start Date</mat-label>
+        <input
+          matInput
+          [matDatepicker]="startPicker"
+          [min]="currentDate"
+          [(ngModel)]="startDate"
+          (dateChange)="onStartDateChange($event)"
+          [disabled]="data.currentState"
+          readonly
+          required
+          (click)="startPicker.open()"
+        />
+        <mat-datepicker-toggle
+          matIconSuffix
+          [for]="startPicker"
+        ></mat-datepicker-toggle>
+        <mat-datepicker #startPicker></mat-datepicker>
+
+        <!-- Simplified logic for Start Date mat-hint -->
+        @if (data.type === 'all_preferences' || data.type ===
+        'single_preferences') {
+        <mat-hint *ngIf="!data.currentState">
+          The submission will start @if (isStartDateToday) {
+          <strong>today, immediately</strong>
+          } @else { in <strong>{{ remainingDaysStart }} day(s)</strong>
+          }
+        </mat-hint>
+        }
+        <!-- no mat-hint if data.currentState === true -->
+      </mat-form-field>
+    </div>
+    }
+
+    <!-- Submission Deadline Date Picker Section -->
+    @if (showDeadlinePicker && !isPreferencesScheduled) {
+    <div class="date-picker">
+      <mat-form-field appearance="fill">
+        <mat-label>End Date</mat-label>
+        <input
+          matInput
+          [matDatepicker]="endPicker"
+          [min]="minDate"
+          [(ngModel)]="submissionDeadline"
+          (dateChange)="onDeadlineChange($event)"
+          [disabled]="data.currentState"
+          readonly
+          required
+          (click)="endPicker.open()"
+        />
+        <mat-datepicker-toggle
+          matIconSuffix
+          [for]="endPicker"
+        ></mat-datepicker-toggle>
+        <mat-datepicker #endPicker></mat-datepicker>
+
+        <!-- Simplified logic for End Date mat-hint -->
+        @if (data.type === 'all_preferences' || data.type ===
+        'single_preferences') {
+        <mat-hint *ngIf="data.currentState">
+          The submission is set to close in @if (isDeadlineToday) {
+          <strong>today at 11:59 PM</strong>
+          } @else {
+          <strong>{{ remainingDays }} day(s)</strong>
+          }
+        </mat-hint>
+        }
+        <!-- no mat-hint if data.currentState === false -->
+      </mat-form-field>
+    </div>
+    }
+
+    <!-- Email Notification Section -->
+    @if (showEmailOption && !data.currentState && !isPreferencesScheduled) {
+    <div class="checkbox-container email">
+      <mat-checkbox [(ngModel)]="sendEmail" class="email-checkbox">
+        <span class="checkbox-content">
+          <span mat-symbol="email"></span>
+          <span>
+            Send email to @if (data.type === 'single_preferences') {
+            {{ data.facultyName }}
+            } @else { all faculty members }
+          </span>
+        </span>
+      </mat-checkbox>
+    </div>
+    }
+
+    <!-- Confirmation Text -->
+    @if ((data.type === 'all_preferences' || data.type === 'single_preferences')
+    && data.currentState) {
+    <div class="main-text">
+      <div class="confirm-disable">
+        This will <strong>immediately</strong> disable the preferences
+        submission and bypass the currently set deadline.
+      </div>
+    </div>
+    } @if (data.type === 'all_preferences' && !data.currentState &&
+    hasIndividualDeadlines) {
+    <div class="main-text">
+      <div class="confirm-disable">
+        This will bypass individual deadlines set to any faculty members.
+      </div>
+    </div>
+    }
+
+    <!-- New Section for Scheduled Preferences Submission -->
+    @if (isPreferencesScheduled) {
+    <div class="scheduled-preferences">
+      <div class="scheduled-pref-icon">
+        <span mat-symbol="hourglass_top"></span>
+      </div>
+      <div class="scheduled-pref-text">
+        Preferences submission is currently scheduled to open from
+        <strong>{{ startDate | date }}</strong> until
+        <strong>{{ submissionDeadline | date }}</strong
+        >.
+      </div>
+    </div>
+    }
+  </div>
+
+  <!-- Dialog Buttons -->
+  <div class="dialog-buttons">
+    <button
+      mat-button
+      class="close"
+      (click)="closeDialog()"
+      [disabled]="isProcessing"
+    >
+      Close
+    </button>
+
+    @if (isPreferencesScheduled) {
+    <button
+      mat-flat-button
+      color="primary"
+      class="action-button"
+      (click)="cancelScheduledSubmission()"
+      [disabled]="isProcessing"
+      [ngClass]="{
+        'disabled-button': isProcessing
+      }"
+    >
+      @if (isProcessing) {
+      <mat-spinner class="custom-spinner" diameter="25"></mat-spinner>
+      } @else { Cancel Scheduled Submission }
+    </button>
+    } @else {
+    <button
+      mat-flat-button
+      color="primary"
+      class="action-button"
+      (click)="confirmAction()"
+      [disabled]="
+        isProcessing ||
+        (showDeadlinePicker && submissionDeadline === null) ||
+        (showStartDatePicker && startDate === null)
+      "
+      [ngClass]="{
+        'disabled-button':
+          isProcessing ||
+          (showDeadlinePicker && submissionDeadline === null) ||
+          (showStartDatePicker && startDate === null)
+      }"
+    >
+      @if (isProcessing) {
+      <mat-spinner class="custom-spinner" diameter="25"></mat-spinner>
+      } @else { Confirm }
+    </button>
+    }
+  </div>
+</div>
