+++ frontend/src/app/shared/dialog-toggle-preferences/dialog-toggle-preferences.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DialogTogglePreferencesComponent } from './dialog-toggle-preferences.component';
+
+describe('DialogTogglePreferencesComponent', () => {
+  let component: DialogTogglePreferencesComponent;
+  let fixture: ComponentFixture<DialogTogglePreferencesComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [DialogTogglePreferencesComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(DialogTogglePreferencesComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
